chore: code hygiene cleanup

+++ src/ngx_http_coraza_header_filter.c
@@ -297,7 +297,7 @@ ngx_http_coraza_resolv_header_last_modified(ngx_http_request_t *r, ngx_str_t nam
     p = ngx_http_time(buf, r->headers_out.last_modified_time);
 
     value.data = buf;
-    value.len = (int)(p-buf);
+    value.len = (size_t) (p - buf);
 
     return ngx_http_coraza_add_response_header(r, ctx, &name, &value);
 }
@@ -562,7 +562,7 @@ ngx_http_coraza_header_filter(ngx_http_request_t *r)
         }
     }
 
-    /* prepare extra paramters for msc_process_response_headers() */
+    /* prepare extra parameters for coraza_process_response_headers() */
     if (r->err_status) {
         status = r->err_status;
     } else {
@@ -661,26 +661,6 @@ ngx_http_coraza_header_filter(ngx_http_request_t *r)
         }
         return ngx_http_filter_finalize_request(r, &ngx_http_coraza_module, ret);
     }
-
-    /*
-     * Proxies will not like this... but it is necessary to unset
-     * the content length in order to manipulate the content of
-     * response body in Coraza.
-     *
-     * This header may arrive at the client before Coraza had
-     * a change to make any modification. That is why it is necessary
-     * to set this to -1 here.
-     *
-     * We need to have some kind of flag the decide if Coraza
-     * will make a modification or not. If not, keep the content and
-     * make the proxy servers happy.
-     *
-     */
-
-    /*
-     * The line below is commented to make the spdy test to work
-     */
-     //r->headers_out.content_length_n = -1;
 
     /*
      * Delay forwarding response headers until the body filter has finished
